Refactoring and some cleaning. Using also autobind
