+++ WebClient/package.json
@@ -7,8 +7,10 @@
   "author": "Triple A Team",
   "license": "",
   "devDependencies": {
-    "autobind-decorator": "1.3.4",
+    "autobind-decorator": "^1.3.4",
     "babel": "6.23.0",
+    "babel-plugin-transform-decorators": "^6.22.0",
+    "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-preset-es2015": "^6.22.0",
     "babel-preset-react": "^6.23.0",
     "babelify": "7.3.0",
